refactor(android): rename package to com.nostudios.bread
